% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{boot2pval Function}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{test}{}
}
\value{
a hitogram of the tsat based on the parameters provided to the
function.
}
\description{
This function two samples x1, x2, a confidence level percentage, the number
of iterations, the difference in the means of the two samples and the method
of test the t.test should conduct as parameters and plots a histogram of the
tstat.
}
\examples{
set.seed(10);x=rnorm(20,mean=4,sd=2);  y=rnorm(20,mean=4,sd=2)
boot2pval(x1=x,x2=y)

}
