% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage}
\format{

}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected concerning home damage from fire
}
\keyword{datasets}
