% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha, cx = 1.5, ...)
}
\arguments{
\item{...}{}
}
\value{
a histogram with the confidence interval
}
\description{
Uses the bootstrap simulation technique to find the confidence interval of
the mean statistic for a given sample.
}
