% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci() Function}
\usage{
myci(x, p)
}
\arguments{
\item{p}{}
}
\value{
a confidence interval with a lower and upper limit.
}
\description{
Takes the size of the sample and the percentage of probability as parameters
in order to produce a confidence interval as required.
}
