% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson Distribution Function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
a  Poisson Probability Distribution barplot
}
\description{
Takes size (n), number of iterations (iter) and lambda as parameters in order
to produce a Poisson Probability Distribution barplot.
}
\examples{

}
