% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{...}{}
}
\value{
the maximum likelihood of the required function
}
\description{
This function takes the parameters and calculates the maximum liklihood
for a binomial, poisson distribution or exponential function as required
and plots it.
}
\examples{

max likelihood function
For repeated sampling from same distribution
}
