% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal distribution function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{value of mean for the normal distribution function.}

\item{sigma}{value of sigma for the normal distribution function}

\item{a}{the upper limit of the interval.}
}
\value{
a normal distribution curve and probability from negative infinity
to (a).
}
\description{
Takes parameters mu, sigma and a to give a normal distribution curve with
area from - infinity to (a).
}
\examples{
myncurve(10, 5, 6)

}
